StrideArrays.jl

Manual Outline

Cascadelake

Tigerlake

Haswell

StrideArrays.@gc_preserve

StrideArrays.@gc_preserve — Macro

@gc_preserve foo(A, B, C)

Apply to a single, non-nested, function call. It will GC.@preserve all the arguments, and substitute suitable arrays with PtrArrays. This has the benefit of potentially allowing statically sized mutable arrays to be both stack allocated, and passed through a non-inlined function boundary.